feat: add unmapped inventory report and decommisioning service

+++ src/main/java/com/telkom/co/ke/almoptics/repository/tbNodeRepository.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Repository;
  * @author jgithu
  */
 @Repository
-public interface tbNodeRepository extends JpaRepository<tbNode, Long> {
+public interface tbNodeRepository extends JpaRepository<tbNode, Integer> {
 
     tbNode findBySerialNumber(String serialNumber);
